SlideDialog now has its own class
