+++ Hestia/app/src/main/java/com/rugged/application/hestia/SingleFragmentActivity.java
@@ -31,6 +31,7 @@ public abstract class SingleFragmentActivity extends AppCompatActivity{
 
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
 
         FragmentManager fm = getSupportFragmentManager();
         Fragment fragment = fm.findFragmentById(R.id.fragment_container);
